requirements.txt
amqp==2.6.0
asgiref==3.3.1
Babel==2.9.1
beautifulsoup4==4.9.1
billiard==3.6.3.0
blessings==1.7
bpython==0.19
celery==4.4.6
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
curtsies==0.3.5
cwcwidth==0.1.4
defusedxml==0.7.1
Django==3.0.5
django-autoslug==1.9.8
django-celery-beat==2.0.0
django-cors-headers==3.2.1
django-countries==7.2.1
django-crum==0.7.7
django-extensions==3.1.3
django-guardian==2.3.0
django-js-asset==1.2.2
django-language-field==0.0.3
django-mptt==0.13.3
django-multiselectfield==0.1.12
django-phonenumber-field==5.2.0
django-randompinfield==0.4.7
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
django-tenants==3.3.2
django-timezone-field==4.1.2
django-uuidfield==0.5.0
Django-Verify-Email==1.0.7
djangorestframework==3.11.0
djangorestframework-guardian==0.3.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
EasyProcess==0.3
future==0.18.2
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
kombu==4.6.11
llvmlite==0.36.0
MarkupSafe==1.1.1
numba==0.53.1
numpy==1.20.2
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==20.9
phonenumbers==8.12.33
Pillow==8.2.0
psycopg2==2.8.6
pycryptodome==3.10.1
pycryptodomex==3.10.1
pydotplus==2.0.2
Pygments==2.8.1
PyJWT==1.7.1
pyotp==2.6.0
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2019.3
PyVirtualDisplay==2.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
selenium==3.141.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
uWSGI==2.0.19.1
vine==1.3.0
